Add parameter to configure field normalization in plugin
I think it's necessary to leave option to disable this normalization. Also for some data splitting field names by '-', for instance, could make more sense than by '_' or ' '. (Probably?) I decided to make '_', ' ', '\t' a default delimiters because i personally never seen a csv or json with field names separated by anything other than that. Not sure if adding a '-' to this set would be helpful or annoying.