Use embedded server for network related tests
Sometimes remote servers won't respond and cause tests to fail. Also, easier to config response if needed

Duplicated EmbeddedServerRunners.kt in two modules, because i couldn't find a way to depend on test source set of root from plugin's modules. As far as i know, it's possible to create  another module with test utilities, but it seems an overkill.