Update RELEASE_CHECK_LIST.md for deprecation updatres
We should update the deprecated methods at the beginning of the new cycle right as the working version changes. This will lead to a lot less confusion while working on the project as you know which methods will be removed. Plus you can introduce new warning deprecations for the upcoming release.