Add JDBC tests notebook
This commit introduces a new notebook 'JDBC tests', testing the connection and integration with PostgreSQL using JDBC. In this notebook, the org.postgresql connection is tested and various DataFrame operations are performed. Essential configurations such as URL, username, and password are also added. It also includes some insights about the expected argument order in the 'readSqlTable' function.