Add support for different IDE versions in DataFrame rendering
This update introduces specific rendering flows based on the version of the IDE. A new method has been added to retrieve the IDE build number. This information is then used to conditionally generate a JSON-encoded DataFrame, adapting to the specific capabilities of the IDE version in use. Additionally, relevant tests have been updated to accommodate these changes.