remove code generation that was needed to glue library and plugin together
The idea was to verify that compile time evaluation and runtime evaluation of the same expressions give the same results.  To run the same code twice, i decided to use code generation. It also was a source of truth about mapping between dataframe function arguments and compile time FIR value that are expected.
Code generation pipeline has proven to be a bit hard to maintain, so approach needs to be changed.