Update SQL all table/schemas reading functions to return maps with table names (#718)
* Update SQL table reading functions to return maps

Refactored the SQL table reading functions to return map data structures instead of list. This change helps to easily correlate each dataframe with its underlying table. Additionally, the function comments and test cases were updated to match this change.

* Correct typos and ignore MSSQLTest

Typographical errors in the comments of the readJdbc.kt file were corrected for clarity. Additionally, the MSSQLTest class in the mssqlTest.kt file was annotated with "@Ignore" to skip the test in subsequent testing runs.